+++ elynx-tree/src/ELynx/Tree/Parallel.hs
@@ -19,7 +19,6 @@
 module ELynx.Tree.Parallel
   ( parTree,
     parBranchFoldMap,
-    parBranchFoldMapWithReversedPath,
     parNodeFoldMap,
   )
 where
@@ -27,7 +26,6 @@ where
 import Control.Parallel.Strategies
 import Data.Foldable
 import ELynx.Tree.Rooted
-import ELynx.Tree.Zipper
 
 myParList :: Strategy a -> Strategy [a]
 myParList _ [] = return []
@@ -57,46 +55,6 @@ parBranchFoldMap :: NFData f => Int -> (e -> f) -> (f -> f -> f) -> Tree e a -> 
 parBranchFoldMap n f op t@(Node br _ ts)
   | n >= 1 = foldl' op (f br) (map (parBranchFoldMap (n - 1) f op) ts `using` myParList rdeepseq)
   | otherwise = branchFoldMap f op t
-
-branchFoldMapWithReversedPath ::
-  -- Layer.
-  Int ->
-  [Int] ->
-  (Path -> e -> f) ->
-  (f -> f -> f) ->
-  Tree e a ->
-  f
-branchFoldMapWithReversedPath d p f op (Node br _ ts) =
-  foldl'
-    op
-    (f p br)
-    (zipWith (\t i -> branchFoldMapWithReversedPath (d + 1) (i : p) f op t) ts [0 ..])
-
--- | Map and fold over branches.
---
--- The used function has access to the __reversed__ path of the node to which
--- the handled branch is attached to. The reversed path is used for reasons of
--- computational speed.
---
--- Evaluate the sub trees up to given layer in parallel.
-parBranchFoldMapWithReversedPath ::
-  NFData f =>
-  Int ->
-  (Path -> e -> f) ->
-  (f -> f -> f) ->
-  Tree e a ->
-  f
-parBranchFoldMapWithReversedPath = go 0 []
-  where
-    go d p n f op tr@(Node br _ ts)
-      | n >= 1 =
-        foldl'
-          op
-          (f p br)
-          ( zipWith (\t i -> go (d + 1) (i : p) (n - 1) f op t) ts [0 ..]
-              `using` myParList rdeepseq
-          )
-      | otherwise = branchFoldMapWithReversedPath d p f op tr
 
 nodeFoldMap :: (a -> b) -> (b -> b -> b) -> Tree e a -> b
 nodeFoldMap f op (Node _ lb ts) = foldl' op (f lb) $ map (nodeFoldMap f op) ts